Fix more tests, fix auto debug name default, even for DxcAssembler
Change to not embed debug with /Zi was even more painful than anticipated.
There are scenarios and code paths that had to be modified to work.
Lots of tests broke in various annoying ways.
Reversed decision to change default debug name to be based on stripped
module instead of debug module (based on source), partly because there
was no way to preserve previous behavior with DxcAssembler.
Instead, default to DebugNameForBinary if no DebugInfo.

Also, since we output debug name if asked for even if we are not
generating debug info, it has to be downgraded from default
DebugNameForSource to DebugNameForBinary when debug info isn't actually
present.  This can happen when in codepath that didn't have options, but we want
to preserve what we have (DxcAssembler).