Update FileCheck Batch tests, sorting to optimize parallel execution
Replaced manually added groups with macro that defines all the groups and paths.
Added missing groups that were in HLSLFileCheck but missing from test.
Broke up the hlsl and samples groups, since these were the long poles.

Added SortFileCheck.py to time and sort the groups for more optimum parallel execution efficiency.
The script also detects when HLSLFileCheck directory is out-of-date with the test groups.