Refactor top-level IDxbcConverter impl to ensure fresh converter state
Previously, if you converted a shader and retained the interface instance,
subsequent shader conversions could be impacted by stale state leftover
from the previous conversion.

This change refactors the top-level implementation of the IDxbcConverter
interface to put the converter implementation in a separate class.

The DxbcConverterFactory class is responsible for creating instances of
the DxbcConverter that actually implements the conversion using the virtual
Create() method.  It also has the virtual LogConvertResult method.

Implementations can create derived versions of the DxbcConverterFactory and
DxbcConverter classes to extend functionality.