Add asan/lsan leak suppression
address-sanitizer reports a leak coming from the dynamic shared library loading code, though dlopen() is paired with dlclose().
This leak doesn't manifest on main yet because dxc adds dxcompiler directly to target_link_libraries (which it shouldn't have to), and dxil is loaded within that library.
A PR to remove dxil loading from dxcompiler exposes the leak when dxc loads dxil dynamically.

This change adds a suppressions file for LSAN to suppress the leak in the loader code that happens under call_init.
To make this work, it must also change the Release build to RelWithDebInfo so symbols are present, which are required to resolve the function name we match in the suppressions file.